Consolidate flex and container into a shared base class `gds-div`
