+++ libs/core/src/components/div/div.style.ts
@@ -1,0 +1,11 @@
+import { css } from 'lit'
+
+const style = css`
+  :host(:focus-visible) {
+    outline-color: color-mix(in srgb, currentcolor, #000 100%);
+    outline-offset: var(--gds-space-3xs);
+    outline-style: solid;
+    outline-width: var(--gds-space-3xs);
+  }
+`
+export default style
