+++ libs/core/src/components/div/index.ts
@@ -1,0 +1,1 @@
+export * from './div'
