+++ libs/core/src/components/flex/flex.ts
@@ -1,11 +1,9 @@
 import { tokens } from '../../tokens.style'
-import { styleExpressionProperty } from '../../utils/decorators/style-expression-property'
-import { forSpaceTokens } from '../../utils/helpers'
 import {
   gdsCustomElement,
   html,
 } from '../../utils/helpers/custom-element-scoping'
-import { GdsContainer } from '../container'
+import { GdsDiv } from '../div'
 import FlexCSS from './flex.styles'
 
 /**
@@ -19,107 +17,8 @@ import FlexCSS from './flex.styles'
  */
 
 @gdsCustomElement('gds-flex')
-export class GdsFlex extends GdsContainer {
+export class GdsFlex extends GdsDiv {
   static styles = [tokens, FlexCSS]
-
-  /**
-   * Controls the gap property of the flex.
-   * Supports all space tokens as value.
-   *
-   * You can apply gap in each column and row like this:
-   * ```html
-   * <gds-flex gap="m s"></gds-flex>
-   * ```
-   *
-   * The `gap` can be applied on `flex` using shorthand like gap="m s"
-   *
-   * Gap also support breakpoint syntax like this:
-   * ```html
-   * <gds-flex gap="xs{s} m{m} l{xl}"></gds-flex>
-   * ```
-   * The above example will apply the gap style of `xs` for `small` devices, `m` for `medium` devices, and `xl` for large devices.
-   *
-   */
-  @styleExpressionProperty({
-    ...forSpaceTokens,
-    styleTemplate: (_prop, values) => {
-      const colGap = values[0]
-      const rowGap = values[1] || colGap
-      return `gap: ${colGap} ${rowGap};`
-    },
-    cacheOverrideKey: 'flex',
-  })
-  gap?: string
-
-  /**
-   * Controls the `flex` property of the flex.
-   * Supports all valid CSS flex property values.
-   */
-  @styleExpressionProperty()
-  flex?: string
-
-  /**
-   * Controls the `align-self` property of the flex.
-   * Supports all valid CSS align-self values.
-   */
-  @styleExpressionProperty()
-  'align-self'?: string
-
-  /**
-   * Controls the `align-items` property of the flex.
-   * Supports all valid CSS align-items values.
-   */
-  @styleExpressionProperty()
-  'align-items'?: string
-
-  /**
-   * Controls the `align-content` property of the flex.
-   * Supports all valid CSS align-content values.
-   */
-  @styleExpressionProperty()
-  'align-content'?: string
-
-  /**
-   * Controls the `justify-content` property of the flex.
-   * Supports all valid CSS justify-content values.
-   */
-  @styleExpressionProperty()
-  'justify-content'?: string
-
-  /**
-   * Controls the `place-content` property of the flex.
-   * Supports all valid CSS place-content values.
-   */
-  @styleExpressionProperty()
-  'place-content'?: string
-
-  /**
-   * Controls the `justify-items` property of the flex.
-   * Supports all valid CSS justify-items values.
-   */
-  @styleExpressionProperty()
-  'justify-items'?: string
-
-  /**
-   * Controls the `justify-self` property of the flex.
-   * Supports all valid CSS justify-self values.
-   */
-  @styleExpressionProperty()
-  'justify-self'?: string
-
-  /**
-   * Controls the flex-direction property of the flex.
-   * Supports all valid CSS flex-direction values.
-   */
-  @styleExpressionProperty()
-  'flex-direction'?: string
-
-  /**
-   * Controls the flex-wrap property of the flex.
-   * Supports all valid CSS flex-wrap values.
-   */
-  @styleExpressionProperty()
-  'flex-wrap'?: string
 
   connectedCallback(): void {
     super.connectedCallback()
